Change in DAS28 by 6-month ACR response

Mean of change in DAS28 by 6-month mutually exclusive ACR response category (ACR < 20, ACR 20 - <50, ACR 50 - <70, and ACR 70+).

acr2das28

Format

A list of one data.table with four columns. Column 1 is ACR response category, column 2 is the point estimate for the mean change in DAS28 at 6 months, and columns 3/4 are lower/upper bounds on the mean change in DAS28. Format is a list to facilitate inclusion of evidence from additional data sources with different patient populations.

Source

Aletaha, Daniel, and Josef S. Smolen. "The simplified disease activity index (SDAI) and clinical disease activity index (CDAI) to monitor patients in standard clinical care." Best practice & research Clinical rheumatology 21.4 (2007): 663-675.

Aletaha, Daniel, et al. "Acute phase reactants add little to composite disease activity indices for rheumatoid arthritis: validation of a clinical activity score." Arthritis research & therapy 7.4 (2005): R796.