Survival analysis parameters for therapy duration by Eular response category

A list of parameters estimated in a parametric survival analysis of treatment duration. Models were fit using flexsurv. Survival curves were scanned from the 2016 NICE HTA and then converted to individual-level patient dat using the reconstruct_ipd function. Survival curves are stratified by patients with moderate and good responses to treatment according to Eular response categories.

dur.eular.mod

dur.eular.good

dur.eular.mod.adj

dur.eular.good.adj

Format

A list of lists. One list for each parametric distribution. Within each distribution the list contains parameters produced by flexsurv on real-line scale (e.g. after log transformation if they are defined as positive). The parameters are:

coef

Coefficients for all parameters including location parameter and ancillary parameter(s)

vcov

Covariance matrix for all parameters including location parameter and ancillary parameter(s)

Source

Stevenson, Matt, et al. "Adalimumab, etanercept, infliximab, certolizumab pegol, golimumab, tocilizumab and abatacept for the treatment of rheumatoid arthritis not previously treated with disease-modifying antirheumatic drugs and after the failure of conventional disease-modifying antirheumatic drugs only: systematic review and economic evaluation." Health Technology Assessment 20.35 (2016): 1-610.