Input data for simulation

Generate data inputs for the IPS.

get_input_data(patdata, x_mort = NULL, x_ttd_all = NULL, x_ttd_da = NULL,
  x_ttd_eular = NULL, model_structures)

Arguments

patdata
Matrix of patient data. Must contain variables generated from sample_pats: 'age' for age, 'haq0' for baseline HAQ, 'male' as a indicator equal to 1 if the patient is male and 0 if female, 'weight' for patient weight, and 'prev_dmards' for number of previous DMARDs.

x_mort
A matrix with each column a variable used to adjust mortality.

x_ttd_all
The design matrix for time to treatment discontinuation representative of all patients (i.e., unstratified).

x_ttd_da
The design matrix for time to treatment discontinuation stratified by disease activity level.

x_ttd_eular
The design matrix for time to treatment discontinuation for each EULAR response category (moderate, good).

model_structures
An object of class model_structures generated from select_model_structures.

Value

A list containing the following data inputs:

n

Number of simulated patients

haq0

A vector of patient HAQ at baseline.

age

A vector of patient age at baseline.

male

A vector of patient gender (1 = male, 0 = female).

prev.dmards

A vector of the number of previous DMARDs

x.mort

Design matrix for mortality adjustment with odds ratios

x.ttd

Design matrix for treatment duration.