Input data for simulation

Generate data inputs for the IPS.

input_data(patdata, vars_mort = "haq0", vars_dur = "1")

Arguments

patdata

Matrix of patient data. Must contain variables generated from sample_pats: 'age' for age, 'haq0' for baseline HAQ, 'male' as a indicator equal to 1 if the patient is male and 0 if female, 'weight' for patient weight, 'dis_dur' for disease duration and 'prev_dmards' for number of previous DMARDs.

vars_mort

A chararacter vector of variables in patdata to be used to adjust 'qx' (using odds ratios) in the lifetables. Use '1' to include an intercept.

vars_dur

A character vector of variables in patdata to be used to predict treatment duration. Use '1' to include an intercept.

Value

A list containing the following data inputs:

n

Number of simulated patients

haq0

A vector of patient HAQ at baseline.

age

A vector of patient age at baseline.

male

A vector of patient gender (1 = male, 0 = female).

dis.dur

A vector of disease duration

prev.dmards

A vector of the number of previous DMARDs

x.mort

Design matrix for mortality adjustment with odds ratios

x.dur

Design matrix for treatment duration.