Mortality hazard ratio for change in HAQ score

Impact of change in HAQ on for mortality from Table 3 in Michaud et al (2012).

mort.hr.haqdif

Format

A data frame with 5 rows and 8 variables. Key variables are month, log of hazard ratio, and standard error of log of hazard ratio.

month

Month of reported hazard ratio (\(month \le 6\), \(6 > month \le 12\), \(12 < month \le 24\), \(24 < month \le 36\), \(month > 36\)).

loghr

Log of hazard ratio for impact of .25 unit HAQ increase on mortality

loghr_se

Standard error of log of hazard ratio for impact of .25 unit HAQ increase on mortality

Source

Michaud, Kaleb, Montserrat Vera-Llonch, and Gerry Oster. "Mortality risk by functional status and health-related quality of life in patients with rheumatoid arthritis." The Journal of rheumatology 39.1 (2012): 54-59.