Mortality odds ratios by patient characteristics

Impact of HAQ on odds ratio for mortality from table 4 in Wolfe et al (2003).

mort.or

Format

A data frame with 1 row and 9 columns:

var

Name of variable

or

Odds ratio without radiographic data

or_se

Standard error of odds ratio without radiographic data

logor

Log odds ratio without radiographic data

logor_se

Standard error of log odds ratio without radiographic data

Source

Wolfe, Frederick, et al. "Predicting mortality in patients with rheumatoid arthritis." Arthritis & Rheumatism 48.6 (2003): 1530-1542.

Details

The standard errors of the log odds ratios are derived from the standard error of the odds ratios. By the delta method, \(var(OR) = exp(\beta)var(\beta)exp(\beta)\) where \(\beta\) is the regression coefficient of the logistic regression. Since \(var(\beta)\) is the variance of the log odds ratio, it follows that \(var[log(OR)] = [var(OR)/exp(\beta)\)]^2.