NMA parameters to ACR response probabilities

Calculate ACR response probabilities from NMA ordered probit parameters.

nma_acrprob(A, z2, z3, delta, rr = 1)

Arguments

A
A vector from the posterior distribution of the probability that ACR response < 20 from cDMARDS.

z2
A vector from the posterior distribution for the ACR 50 cutpoint.

z3
A vector from the posterior distribution for the ACR 70 cutpoint.

delta
A matrix from the posterior distribution of regression coefficients (on the probit scale) for each therapy in the NMA.

rr
Sampled value of elative risk.

Value

List containing an array non.overlap (the probability of ACR < 20, ACR 20-50, ACR 50-70, and ACR 70+) and overlap (the probability of ACR20/50/70).