Sample patients

Sample patients for rheumatoid arthritis individual patient simulation.

sample_pats(n = 1, type = c("homog", "heterog"), age_mean = 55,
  age_sd = 13, male_prop = 0.21, haq0_mean = 1.5, haq0_sd = 0.7,
  wtmale = 89, wtfemale = 75, prev_dmards_mean = 3.28,
  prev_dmards_sd = 1.72, das28_mean = 6, das28_sd = 1.2, sdai_mean = 43,
  sdai_sd = 13, cdai_mean = 41, cdai_sd = 13, cor_das28_sdai = 0.86,
  cor_das28_cdai = 0.86, cor_das28_haq = 0.38, cor_sdai_cdai = 0.94,
  cor_sdai_haq = 0.34, cor_cdai_haq = 0.34)

Arguments

n
Number of samples.

type
Should male and female patients be heterogeneous or homogeneous. Default is homogeneous.

age_mean
Mean age.

age_sd
Standard deviation of age.

male_prop
Proportion male.

haq0_mean
Mean baseline HAQ score.

haq0_sd
Standard deviation of baseline HAQ score.

wtmale
Male weight.

wtfemale
Female weight.

prev_dmards_mean
Mean number of previous DMARDs.

prev_dmards_sd
Standard deviation of number of previous DMARDs.

das28_mean
Mean of DAS28.

das28_sd
Standard deviation of DAS28.

sdai_mean
Mean of SDAI.

sdai_sd
Standard deviation of SDAI.

cdai_mean
Mean of CDAI.

cdai_sd
Standard deviation of CDAI.

cor_das28_sdai
Correlation between DAS28 and SDAI.

cor_das28_cdai
Correlation between DAS28 and CDAI.

cor_das28_haq
Correlation between DAS28 and baseline HAQ.

cor_sdai_cdai
Correlation between SDAI and CDAI.

cor_sdai_haq
Correlation between SDAI and HAQ.

cor_cdai_haq
Correlation between CDAI and HAQ.

Value

Matrix of patient characteristics. One row for each patient and one column for each variable. Current variables are:

age

age in years

male

1 = male, 0 = female

weight

Patient weight

prev_dmards

Number of previous DMARDs

das28

DAS28 score

sdai

SDAI score

cdai

CDAI score

haq0

Baseline HAQ score