Sample a patient population

Sample a patient population for use in the individual patient simulation (sim_iviRA).

sample_pop(n = 1, type = c("homog", "heterog"), age_mean = 55,
  age_sd = 13, male_prop = 0.21, haq0_mean = 1.5, haq0_sd = 0.7,
  wtmale = 89, wtfemale = 75, prev_dmards_mean = 3.28,
  prev_dmards_sd = 1.72, das28_mean = 6, das28_sd = 1.2, sdai_mean = 43,
  sdai_sd = 13, cdai_mean = 41, cdai_sd = 13, cor_das28_sdai = 0.86,
  cor_das28_cdai = 0.86, cor_das28_haq = 0.38, cor_sdai_cdai = 0.94,
  cor_sdai_haq = 0.34, cor_cdai_haq = 0.34)

Arguments

n
Number of samples.

type
Should male and female patients be heterogeneous or homogeneous. Default is homogeneous.

age_mean
Mean age.

age_sd
Standard deviation of age.

male_prop
Proportion male.

haq0_mean
Mean baseline HAQ (i.e., HAQ at the start of the model) score.

haq0_sd
Standard deviation of baseline HAQ score.

wtmale
Male weight.

wtfemale
Female weight.

prev_dmards_mean
Mean number of previous DMARDs.

prev_dmards_sd
Standard deviation of number of previous DMARDs.

das28_mean
Mean of DAS28.

das28_sd
Standard deviation of DAS28.

sdai_mean
Mean of SDAI.

sdai_sd
Standard deviation of SDAI.

cdai_mean
Mean of CDAI.

cdai_sd
Standard deviation of CDAI.

cor_das28_sdai
Correlation between DAS28 and SDAI.

cor_das28_cdai
Correlation between DAS28 and CDAI.

cor_das28_haq
Correlation between DAS28 and baseline HAQ.

cor_sdai_cdai
Correlation between SDAI and CDAI.

cor_sdai_haq
Correlation between SDAI and HAQ.

cor_cdai_haq
Correlation between CDAI and HAQ.

Value

Matrix of patient characteristics. One row for each patient and one column for each variable. Current variables are:

age

Age in years.

male

1 = male, 0 = female.

weight

Patient weight in KG.

prev_dmards

Number of previous DMARDs.

das28

DAS28 score.

sdai

SDAI score.

cdai

CDAI score.

haq0

Baseline HAQ score.

Examples

sample_pop(n = 10, type = "heterog", age_mean = 50)

#>            age male weight prev_dmards    das28     sdai     cdai       haq0
#>  [1,] 27.15680    0     75           2 7.775938 56.65645 58.73401 2.41916910
#>  [2,] 69.26645    0     75           5 5.121937 37.62798 31.56455 1.31145364
#>  [3,] 50.46303    1     89           4 5.105261 38.87250 35.96980 1.49617308
#>  [4,] 47.02475    0     75           2 7.337008 52.61008 51.64376 1.16185901
#>  [5,] 35.05341    0     75           5 5.054066 35.23737 35.07829 0.04472744
#>  [6,] 53.47048    0     75           4 7.385975 55.93197 58.01044 0.81204246
#>  [7,] 58.55695    0     75           2 6.176493 34.83110 37.34499 1.35270951
#>  [8,] 65.18576    0     75           3 4.934270 36.74544 33.78261 2.40322100
#>  [9,] 36.51517    1     89           5 5.902747 37.97200 29.12129 0.94576462
#> [10,] 36.35420    0     75           3 7.349944 69.06357 62.74117 2.70328336