Select model structures

Select the model structures to use in the IVI-RA individual patient simulation.

select_model_structures(tx_ihaq = "acr-haq", tx_iswitch = "acr-switch",
  cdmards_haq_model = "lcgm", ttd_cause = "all", ttd_dist = "exponential",
  utility_model = "mixture")

Arguments

tx_ihaq
Model structure relating treatment to HAQ during the first 6 months of treatment. Options, which are equivalent to H1-H3 in the documentation are:

acr-haq H1: Treatment -> ACR -> HAQ

acr-eular-haq H2: Treatment -> ACR -> EULAR -> HAQ

haq H3:Treatment -> HAQ

tx_iswitch
Model structure relating treatment to switching during the first 6 months of treatment. Options, which are equivalent to S1-S6 in the documentation are:

acr-switch S1: Treatment -> ACR -> Switch

acr-das28-switch S2: Treatment -> ACR -> DAS28 -> Switch

acr-sdai-switch S3: Treatment -> ACR -> SDAI -> Switch

acr-cdai-switch S4: Treatment -> ACR -> CDAI -> Switch

das28-switch S5: Treatment -> DAS28 -> Switch

acr-eular-switch S6: Treatment -> ACR -> EULAR -> Switch

cdmards_haq_model
Model used for long-term HAQ progression. Options are:

lcgm Latent class growth model

linear Constant linear HAQ progression

If lgcm is chosen, then a latent class growth model is used for cDMARDs and NBT but a constant annual rate is is assumed for all other therapies; otherwise a constant linear HAQ progression is assumed for all therapies including cDMARDs and NBT.

ttd_cause
Cause of treatment discontinuation. Options are:

all Treatment discontinuation due to any cause.

si Treatment discontinuation due to serious infections

ttd_dist
Distribution used to model time to treatment discontinuaton. Options are:

exponential Exponential

weibull Weibull

gompertz Gompertz

gamma Gamma

llogis Log-logistic

lnorm Lognormal

gengamma Generalized gamma

utility_model
Model used to estimate patient utility as a function of HAQ and patient characteristics. Options are:

mixture Hernandez Alava (2013) mixutre model

wailoo Wailoo 2006 logistc regression