Simulate change in HAQ at 6 months

Simulate change in HAQ at 6 months using different model structures relating treatment to HAQ.

sim_dhaq6(treatments, input_data, pars, tx_lookup = iviRA::treatments$sname,
  hist = c("naive", "experienced"), line = 1, tx_ihaq = c("acr-haq",
  "acr-eular-haq", "haq"))

Arguments

treatments
Vector of treatments arms to simulate. Should correpond to sname in iviRA::treatments.

input_data
An object of class input_data returned from get_input_data. The only elements required are x_acr and x_haq.

pars
List of sampled parameter values generated from get_input_data. The only elements required are acr, haq, acr2haq, acr2eular, and eular2haq.

tx_lookup
Vector of treatments with names equivalent to iviRA::treatments$sname. Index of treatments in treatments are matched against treatments in tx_lookup by name. Indices of parameter estimates from the network meta-analyses bust be in the same order as in tx_data.

hist
Is the patient biologic naive or biologic experienced?

line
Line of therapy

tx_ihaq
Equivalent to argument tx_haq in select_model_structures.

Value

Change in HAQ for each selected model structure, randomly sampled parameter set, simulated patient, and treatment.