Simulate health care sector costs from simulated HAQ score

Simulate health care sector costs from HAQ score simulated using sim_haq

sim_hc_cost(simhaq, weight, pars,
  cdmards.ind = which(therapy.pars$info[["sname"]] == "cdmards"),
  tcz.ind = which(therapy.pars$info[["sname"]] == "tcz"),
  tczmtx.ind = which(therapy.pars$info[["sname"]] == "tczmtx"),
  check = TRUE)

Arguments

simhaq
Simulation output from sim_haq.

weight
Weight of patient. Used to calculate treatment costs when dosing is based on weight.

pars
List of sampled parameter values. The format is the same as the ouptut produced with sample_pars. The list must contain:

treat.cost

Matrix of treatment cost data.

hosp.cost

List of two matrices used to simulate hospital costs.

mgmt.cost

Matrix of general management costs.

cdmards.ind
Index for cDMARDs.

tcz.ind
Index for tocilizumab.

tczmtx.ind
Index for tocilizuma + methotrexate.

check
Should the function check parameters and input data passed to the model? Default is TRUE.

Value

Matrix. First column is simulated pain score and second column is simulated utility.