Simulate utility using Hernandez-Alava mixture model

Simulate utility from HAQ score simulated using using mixture model from Hernandez-Alva (2013). Can be used to simulate utility after running ivi_RA.

sim_utility_mixture(simhaq, male, pars, check = TRUE)

Arguments

simhaq
Simulation output from sim_haq.

male
Vector indiciating patient gender (1 = male, 0 = female).

pars
List of parameters needed to simulate utility using the Hernandez (2013) mixture model. These are the parameters 'utility.mixture' described in the documentation to sample_pars.

check
Should the function check parameters and input data passed to the model? Default is TRUE.

Value

Matrix. First column is simulated pain score and second column is simulated utility. Each row corresponds to a unique patient and time-period (i.e. month) from sim_haq.