Therapy parameters

Parameters for effect of therapy on 6 month ACR response and haq progression rates. Also includes parameters needed to estimate treatment costs.

therapy.pars

Format

A list of the following lists

info

Basic information about therapies in model

icon_acr_nma

Effect of treatment on ACR response probability from the icon NMA. Contains mean/sd for parameters in the NMA (Baseline effect for cDMARDs A, cutpoints z, and treatment effects delta). Also contains predicted probability of ACR category with parameters at posterior means.

nice_acr_nma

Effect of treatment on ACR response probability from the NICE NMA. Contains probability of each ACR response category by therapy as well as estimated number of observations needed to draw from a Dirichlet distribution.

haq_prog

Effect of treatment on HAQ progression rate. In particular, contains mean and standard error of yearly change in HAQ progression by therapy

cost

Treatment cost by therapy.

Source

nice_acr_nma

Table 37, Stevenson, Matt, et al. "Adalimumab, etanercept, infliximab, certolizumab pegol, golimumab, tocilizumab and abatacept for the treatment of rheumatoid arthritis not previously treated with disease-modifying antirheumatic drugs and after the failure of conventional disease-modifying antirheumatic drugs only: systematic review and economic evaluation." Health Technology Assessment 20.35 (2016): 1-610.

haq_prog

Wolfe, Frederick, and Kaleb Michaud. "The loss of health status in rheumatoid arthritis and the effect of biologic therapy: a longitudinal observational study." Arthritis research & therapy 12.2 (2010): 1.