Treatment switching at 6 months

Regression coefficients from a logistic regression equation predicting the probability of switching treatment at 6 months by disease activity level.

treat.switch

Format

A data.table containing the following columns:

variable

Name of variable. Includes the intercept, an indicator variable for whether disease activity is moderate (CDAI 10-22), and an indicator variable for whether disease activity is high (CDAI greater than 22).

or

The odds ratio of the coefficients.

or_lower

Lower 95 percent confidence interval for the odds ratio.

or_upper

Upper 95 percent confidence interval for the odds ratio.

logor

Log odds ratio of the coefficents.

logor_lower

Lower 95 percent confidence interval for the log of the odds ratio.

logor_upper

Upper 95 percent confidence interval for the log of the odds ratio.

Source

Zhang, Jie, et al. "Thresholds in disease activity for switching biologics in rheumatoid arthritis patients: experience from a large US cohort." Arthritis care & research 63.12 (2011): 1672-1679.