Time to treatment discontinuation parameters

ttd.all, ttd.da, and ttd.eular are lists of lists containing parameters from statistical models of time to treatment discontinuation. ttd.all contains parameters representative of all patients, ttd.da contains covariates for disease activity level (moderate, high), and ttd.eular is stratified by EULAR response (moderate, good).

ttd.eular

ttd.da

ttd.all

Format

A list as described in sample_pars

Source

Stevenson, Matt, et al. "Adalimumab, etanercept, infliximab, certolizumab pegol, golimumab, tocilizumab and abatacept for the treatment of rheumatoid arthritis not previously treated with disease-modifying antirheumatic drugs and after the failure of conventional disease-modifying antirheumatic drugs only: systematic review and economic evaluation." Health Technology Assessment 20.35 (2016): 1-610.

Zhang, Jie, et al. "Thresholds in disease activity for switching biologics in rheumatoid arthritis patients: experience from a large US cohort." Arthritis care & research 63.12 (2011): 1672-1679.

Guyot, Patricia, et al. "Enhanced secondary analysis of survival data: reconstructing the data from published Kaplan-Meier survival curves." BMC medical research methodology 12.1 (2012): 9.

Details

Models were fit using flexsurv. Survival curves are based on analyses of reconstructed individual patient data.