Treatment costs

Acquisition and administration costs for medications used to treat rheumatoid arthritis.

tx.cost

Format

A list with two elements: cost and lookup. The element cost is a data table of variables needed to calculate acquisition and administration costs. Each row denotes a unique treatment. There are 17 variables:

name

Long-form name of treatment.

sname

Short-form name of treatment.

dosage

FDA approved dosage.

strength_dosage_form

Strength and dosage form of treatment

init_dose_val

Numeric value of dose approved by FDA during the first 6 months of treatment.

ann_dose_val

Numeric value of dose approved by FDA beyond the first 6 months.

dose_unit

Unit of FDA approved dose.

init_num_doses

Number of doses based on FDA used during the first 6 months.

ann_num_doses

Annual number of doses based on FDA beyond the first 6 months.

strength_val

Numeric value of strength of treatment.

strength_unit

Unit of treatment.

price_per_unit

Price per unit of treatment. By default, wholesale acquisition costs (WACs) are used.

discount_lower

Discounts and rebates negotiated with manufacturers typically decrease prices paid by patients. Specificy a lower bound for this discount as a proportion of price_per_unit.

discount_upper

An upper bound for the discount as a proportion of price_per_unit.

infusion_cost

Cost per infusion.

loading_dose

Equal to 1 if a loading dose is used and 0 otherwise. If a loading dose is used, then the infusion cost is only applied to the first dose.

weight_based

If weight_based is equal to 1, then dosing is based on each patient's weight (i.e., as sampled by sample_pop).

The element lookup is a data table with 3 variables. The first variable, sname, is the short-form name of a treatment (e.g., corresponding to sname in iviRA::treatments). All remaining columns are the unique agents making up a given treatment. For example, abatecept IV + methotrexate consists of abatecept IV and methotrexate.