Sample parameters

Randomly draw parameters from their (joint) probability distribution.

sample_pars(n = 100, rebound_lower = 0.7, rebound_upper = 1,
  ltfemale = lifetable.female, ltmale = lifetable.male,
  acr2eular_mat = acr2eular, treat_cost = therapy.pars$cost,
  mort_logor = mort.or$logor, mort_logor_se = mort.or$logor_se,
  mort_loghr_haqdif = mort.hr.haqdif$loghr,
  mort_loghr_se_haqdif = mort.hr.haqdif$loghr_se,
  dur_eular_mod = dur.eular.mod.adj, dur_eular_good = dur.eular.good.adj,
  nma1_mean = therapy.pars$icon_acr_nma_naive$mean,
  nma1_sd = sqrt(diag(therapy.pars$icon_acr_nma_naive$vcov)),
  nma2_mean = NULL, nma2_sd = NULL, basedif_mean = 0.25,
  basedif_sd = 0.05, zdif_mean = c(0.01, 0.09), zdif_sd = c(0.05, 0.05),
  treat_hist = c("naive", "exp"),
  haq_lprog_therapy_mean = therapy.pars$haq.lprog$est,
  haq_lprog_therapy_se = therapy.pars$haq.lprog$se,
  eular2haq_mean = eular2haq$mean, eular2haq_se = eular2haq$se,
  acr2haq_mean = acr2haq$mean, acr2haq_se = acr2haq$se,
  acr2sdai_lower = acr2sdai$inception$lower,
  acr2sdai_upper = acr2sdai$inception$upper,
  acr2cdai_lower = acr2cdai$inception$lower,
  acr2cdai_upper = acr2cdai$inception$upper,
  acr2das28_lower = acr2das28$inception$lower,
  acr2das28_upper = acr2das28$inception$upper,
  haq_lprog_age_mean = haq.lprog.age$est,
  haq_lprog_age_se = haq.lprog.age$se, hosp_days_mean = c(0.26, 0.13, 0.51,
  0.72, 1.86, 4.16), hosp_days_se = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5),
  hosp_cost_mean = rep(1251, 6), hosp_cost_se = rep(191, 6),
  mgmt_cost_mean = mgmt.cost$est, mgmt_cost_se = mgmt.cost$se,
  pl_mean = 5853.3, pl_se = 1526.691, si_surv = therapy.pars$si,
  si_cost = 5873, si_cost_range = 0.2, si_ul = 0.156, si_ul_range = 0.2,
  therapy_names = therapy.pars$info$sname)

Arguments

n
Size of the posterior sample.

rebound_lower
The rebound is the increase in HAQ following treatment. It is defined as a proportion f times the size of the inititial treatment response. rebound_lower defines the lower bound for f. Default is 0.7, which implies that the rebound post treatment is 0.7 times the initial treatment effect.

rebound_upper
rebound_upper defines the upper bound for f. Default is 1, which implies that the rebound post treatment is the same as the initial treatment effect.

ltfemale
Lifetable for women. Must contain column 'age' for single-year of age and 'qx' for the probability of death at a given age.

ltmale
Identical to ltfemale but for men.

acr2eular_mat
A two-way frequency matrix with columns denoting EULAR response (none, moderate, good) and rows denoting ACR response (<20, 20-50, 50-70, 70+).

treat_cost
Treatment cost matrix in format of therapy.pars$cost.

mort_logor
Log odds ratio of impact of baseline HAQ on probability of mortality.

mort_logor_se
Standard error of log odds ratio of impact of baseline HAQ on probability of mortality.

mort_loghr_haqdif
Log hazard ratio of impact of change in HAQ from baseline on mortality rate. A vector with each element denoting (in order) hazard ratio for months 0-6, >6 - 12, >12 - 24, >24 -36, >36.

mort_loghr_se_haqdif
Standard error of log hazard ratio of impact of change in HAQ from baseline on mortality rate.

dur_eular_mod
A list containing treatment duration parameters for patientes with a moderate EULAR response. See 'Treatment duration'.

dur_eular_good
A list containing treatment duration parameters for patientes with a good EULAR response. See 'Treatment duration'.

nma1_mean
Posterior means for NMA parameters on probit scale (1st line). NMA estimtes impact of therapy on ACR response.

nma1_sd
Posterior sd for NMA paramters on probit scale (1st line). Used as diagonal for variance-covariance matrix.

nma2_mean
Posterior means for NMA parameters on probit scale (2+ lines). NMA estimtes impact of therapy on ACR response.

nma2_sd
Posterior sd for NMA paramters on probit scale (2+ lines). Used as diagonal for variance-covariance matrix.

basedif_mean
Difference of mean effect of baseline treatment on probit scale between 2nd line and 1st line. Used if there are no 2nd line NMA results.

basedif_sd
Standard deviation of basedif_mean

zdif_mean
Difference of ACR 50 and ACR 70 cutpoints on probit scale between 2nd line and 1st line.

treat_hist
Is patient TIM naive or experienced. If naive NMA results for TIM naive patients are used during 1st line and TIM experienced NMA results are used for subsequent lines. If experienced, NMA results for TIM experienced patients are using during 1st and subsequent lines.

haq_lprog_therapy_mean
Point estimate of linear yearly HAQ progression rate by therapy.

haq_lprog_therapy_se
Standard error of linear yearly HAQ progression rate by therapy.

eular2haq_mean
Mean HAQ change by Eular response category.

eular2haq_se
Standard error of mean HAQ change by Eular response category.

acr2haq_mean
Mean HAQ change by ACR response category.

acr2haq_se
Standard error of mean HAQ change by ACR response category.

acr2sdai_lower
Lower bound for change in SDAI by ACR response category.

acr2sdai_upper
Upper bound for change in SDAI by ACR response category.

acr2cdai_lower
Lower bound for change in CDAI by ACR response category.

acr2cdai_upper
Upper bound for change in CDAI by ACR response category.

acr2das28_lower
Lower bound for change in DAS28 by ACR response category.

acr2das28_upper
Upper bound for change in DAS28 by ACR response category.

haq_lprog_age_mean
Impact of age on annual linear haq progression rate.

haq_lprog_age_se
Standard error of impact of age on annual linear haq progression rate.

hosp_days_mean
Vector denoting average number of hospital days for HAQ < 0.5, 0.5 <= HAQ < 1, 1 <= HAQ < 1.5, 1.5 <= HAQ < 2, 2 <= HAQ < 2.5, HAQ >= 2.5.

hosp_days_se
Vector denoting standard error of average number of hospital days for HAQ < 0.5, 0.5 <= HAQ < 1, 1 <= HAQ < 1.5, 1.5 <= HAQ < 2, 2 <= HAQ < 2.5, HAQ >= 2.5.

hosp_cost_mean
Mean of daily hospital cost.

hosp_cost_se
Standard error of dail hospital cost.

mgmt_cost_mean
Mean of costs of services (in order: chest x-ray, x-ray visit, outpatient followup, Mantoux tuberculin skin test) for general management of RA.

mgmt_cost_se
Standard error of mean of costs of services for general management of RA for general management of RA.

pl_mean
Mean annual productivity loss per 1-unit increase in HAQ.

pl_se
Standard error of mean annual productivity loss per 1-unit increase in HAQ.

si_surv
Paramters of survival model used to estimate time to serious infection.

si_cost
Cost of a serious infection.

si_cost_range
Range used to vary serious infection cost. Default is to calculate upper and lower bound by multiplying si_cost by 1 +/- 0.2 (i.e. a 20% change).

si_ul
One month loss in utility from a serious infection.

si_ul_range
Range used to vary serious infection utility loss. Default is to calculate upper and lower bound by multiplying si_ul by 1 +/- 0.2 (i.e. a 20% change).

therapy_names
Vector of therapy names.

Value

List containing samples for the following model parameters:

treat.cost

data.table of treatment cost data.

ltmale

Lifetable for males with qx transformed using the logit function.

ltfemale

Lifetable for females with qx transformed using the logit function.

acr2eular

An array of matrices. Each matrix represents a random sample of the conditional probability of each EULAR response category for a given ACR response.

logor.mort

Matrix of log odds ratio used to adjust mortality. One row for each sample and one column for each variable used to adjust mortality.

mort.loghr.haqdif

Matrix of the log hazard ratio of the impact of a change in HAQ from baseline on mortality. Columns denote hazard ratios at times < 6 months, months 6 - <12, months 12 - <24, months 24 - <36, and months 36+.

dur.eular.mod

Matrix of coefficients from a survival model. From a survival model for moderate Eular responders

dur.eular.good

Matrix of coefficients for the location parameter and a vector of sampled values of the ancillary parameter. From a survival model for good Eular responders.

acr1

List of two matrices p and po with each row a sampled parameter value. The four columns in p are mutually exclusive categories (ACR <20, ACR 20-50, ACR 50-70, and ACR 70+) and the four columns in po are overlapping categories (ACR < 20, ACR 20, ACR 50, and ACR 70

.
acr2

Subject to change.

eular2haq

A matrix of sampled HAQ changes by Eular response category. The matrix has three columns for no response, moderate response, and good response.

acr2haq

A matrix of sampled HAQ changes by ACR response category. The matrix has four columns for ACR < 20, ACR 20-50, ACR 50-70, and ACR 70+.

haq.lprog.therapy

A matrix of sampled yearly linear change in HAQ by therapy. The matrix has one column for each therapy in therapy.pars.

haq.lprog.age

A matrix of sampled yearly linear change in HAQ by age. The matrix has three columns for age < 40, age 40-64, and age 65+.

mixture.utility

A list containing samples of all parameters in the Hernandez Alva (2013) mixture model. See 'Sampled mixture model parameters' for details.

wailoo.utility

A matrix of sampled regression coefficients from the model mapping HAQ to EQ5D utility in Wailoo (2006). Variables are (in order) "int" (intercept), "age" (patient age), "dis_dur" (disease duration), "haq0" (baseline HAQ), "male" (1 = male, 0 = female), "prev_dmards" (number of previous DMARDs), and "haq" (current HAQ).

hosp.cost

A list of two matrices hosp.days and cost.pday. hosp.days is sample of hospital days by HAQ category; the columns of the matrix are the six HAQ categories (HAQ < 0.5, 0.5 <= HAQ < 1, 1 <= HAQ < 1.5, 1.5 <= HAQ < 2, 2 <= HAQ < 2.5, HAQ >= 2.5). in hosp.days are HAQ. cost.pday is a sample of the costs per hospital day by HAQ category; the columns are the same six HAQ categories as in hosp.days.

mgmt.cost

Matrix of sampled values of general management costs. Each column is a different category of costs ( chest x-ray, x-ray visit, outpatient follow-up, and Mantoux tuberculin skin test).

prod.loss

Vector of sampled values of decrease in wages (e.g. productivity loss) per unit increase in HAQ.

si.surv

Matrix of coefficients from a survival model. Columns numbers coincide with therapy indices.

si.cost

Vector of sampled values of the medical cost of a serious infection.

si.ul

Vector of the sampled values of the annualized utility loss from a serious infection.

Treatment duration

Treatment duration parameters should be contained in a list of lists. The top-level list identifies the name of the probability distribution; the possible distributions are the exponential (exponential), Weibull (weibull), Gompertz (gompertz), gamma (gamma), log-logistic (llogis), lognormal (lnorm), and generalized gamma (gengamma). Each distribution should also contain a list with five elements:

est

A vector of the maximum likelihood estimates of the parameters.

vcov

The variance-covariance of the parameters.

loc.index

A vector of the indices of the location parameters.

anc1.index

A vector of the indices of the first ancillary parameter.

anc2.index

A vector of indices of the second ancillary parameter.

The maximum likelihood estimates should be transformed to the real line. For example, if the model is fit using flexsurvreg in the flexsurv package, the output should be returned from res.t.

Sampled mixture model parameters

The sampled mixture model parameters are contained in a list containing the following:

beta1

Coefficients for class 1 explanatory variables. A matrix of random draws where each column is an explanatory variable.

beta2

Coefficients for class 2 explanatory variables. A matrix of random draws where each column is an explanatory variable.

beta3

Coefficients for class 3 explanatory variables. A matrix of random draws where each column is an explanatory variable.

beta4

Coefficients for class 4 explanatory variables. A matrix of random draws where each column is an explanatory variable.

alpha1

Random effects intecept term for class 1. A vector of random draws.

alpha2

Random effects intecept term for class 2. A vector of random draws.

alpha3

Random effects intecept term for class 3. A vector of random draws.

alpha4

Random effects intecept term for class 4. A vector of random draws.

alpha

Random effects term for male indicator variable. A vector of random draws.

epsilon1

Variance for class 1. A vector of random draws.

epsilon2

Variance for class 2. A vector of random draws.

epsilon3

Variance for class 3. A vector of random draws.

epsilon4

Variance for class 4. A vector of random draws.

mu

Random effects variance term. A vector of random draws.

delta

Coefficients for explanatory variables explaining the probbaility of class membership. An array of matrices where each matrix is a random draw. There are three rows in each matrix (one for each class) and four columns (one for each explanatory variable).