Sample a patient population

Sample a patient population for use in the individual patient simulation (sim_iviRA).

sample_pop(n = 1, type = c("homog", "heterog"), age_mean = 55,
  age_sd = 13, male_prop = 0.21, haq0_mean = 1.5, haq0_sd = 0.7,
  wtmale = 89, wtfemale = 75, prev_dmards_mean = 3.28,
  prev_dmards_sd = 1.72, das28_mean = 6, das28_sd = 1.2,
  sdai_mean = 43, sdai_sd = 13, cdai_mean = 41, cdai_sd = 13,
  cor_das28_sdai = 0.86, cor_das28_cdai = 0.86, cor_das28_haq = 0.38,
  cor_sdai_cdai = 0.94, cor_sdai_haq = 0.34, cor_cdai_haq = 0.34)

Arguments

n
Number of samples.

type
Should male and female patients be heterogeneous or homogeneous. Default is homogeneous.

age_mean
Mean age.

age_sd
Standard deviation of age.

male_prop
Proportion male.

haq0_mean
Mean baseline HAQ (i.e., HAQ at the start of the model) score.

haq0_sd
Standard deviation of baseline HAQ score.

wtmale
Male weight.

wtfemale
Female weight.

prev_dmards_mean
Mean number of previous DMARDs.

prev_dmards_sd
Standard deviation of number of previous DMARDs.

das28_mean
Mean of DAS28.

das28_sd
Standard deviation of DAS28.

sdai_mean
Mean of SDAI.

sdai_sd
Standard deviation of SDAI.

cdai_mean
Mean of CDAI.

cdai_sd
Standard deviation of CDAI.

cor_das28_sdai
Correlation between DAS28 and SDAI.

cor_das28_cdai
Correlation between DAS28 and CDAI.

cor_das28_haq
Correlation between DAS28 and baseline HAQ.

cor_sdai_cdai
Correlation between SDAI and CDAI.

cor_sdai_haq
Correlation between SDAI and HAQ.

cor_cdai_haq
Correlation between CDAI and HAQ.

Value

Matrix of patient characteristics. One row for each patient and one column for each variable. Current variables are:

age

Age in years.

male

1 = male, 0 = female.

weight

Patient weight in KG.

prev_dmards

Number of previous DMARDs.

das28

DAS28 score.

sdai

SDAI score.

cdai

CDAI score.

haq0

Baseline HAQ score.

Examples

sample_pop(n = 10, type = "heterog", age_mean = 50)

#>            age male weight prev_dmards    das28     sdai     cdai      haq0
#>  [1,] 57.80100    1     89           3 7.634467 65.35644 62.76758 1.2817643
#>  [2,] 45.49448    0     75           5 6.376763 50.67509 44.82911 2.0146699
#>  [3,] 46.29300    0     75           4 6.998589 43.42228 48.90367 1.2222298
#>  [4,] 41.58052    0     75           3 3.603166 23.54906 18.04258 1.2235002
#>  [5,] 33.46121    0     75           1 7.674845 57.42469 55.57379 0.7902820
#>  [6,] 47.72275    0     75           4 6.395844 54.80798 54.77467 2.2995913
#>  [7,] 21.38285    0     75           5 5.549343 33.93788 29.69137 0.7628540
#>  [8,] 57.53839    0     75           4 7.027274 50.28449 58.01731 2.7244855
#>  [9,] 50.81635    0     75           4 5.079985 48.19259 41.99686 0.6317067
#> [10,] 73.07662    0     75           5 6.645056 46.45549 37.84887 0.8588147