Simulate HAQ over time

An individual patient simulation of HAQ scores for rheumatoid arthritis patients given a sequence of J treatments. An "outer" loop iterates over S draws from sampled parameter sets and an "inner" loop iterates of N patients. Cycle lengths are 6 months long. The simulation is written in C++ for speed.

sim_haq(arminds, haq0, das28_0, sdai0, cdai0, age0, male, prev_dmards,
  itreat_haq, itreat_switch, nma_acr1, nma_acr2, nma_dhaq1, nma_dhaq2,
  nma_das28_1, nma_das28_2, acr2eular, acr2haq, eular2haq, acr2das28, acr2sdai,
  acr2cdai, tswitch_da, haq_lprog_therapy, haq_lprog_age, haq_lcgm_delta,
  haq_lcgm_beta, cmdards_haq_model, rebound_factor, lifetable_male,
  lifetable_female, x_mort, logor, dur_dist, x_dur, ttd_all_loc, ttd_all_anc1,
  ttd_all_anc2, ttd_da_loc, ttd_da_nac1, ttd_da_anc2, ttd_eular_loc_mod,
  ttd_eular_anc1_mod, ttd_eular_anc2_mod, ttd_eular_loc_good,
  ttd_eular_anc1_good, ttd_eular_anc2_good, cycle_length, treat_gap, cdmards,
  nbt, si_loc, si_anc1, si_anc2, si_dist, haqdelta_loghr, max_months)

Arguments

arminds
Indices of treatment arms consisting of sequences of therapies (each element is the index of a therapy in therapy.pars$info. May be a vector consisting of a single treatment sequence or a matrix of unique sequences for each patient.

itreat_haq
How should the relationship between treatment and HAQ during the initial treatment phase (i.e., first 6 months) be modeled? Options, which are equivalent to H1-H3 in the documentation are H1: Treatment -> ACR -> HAQ (acr-haq), H2: Treatment -> ACR -> EULAR -> HAQ (acr-eular-haq), or H3:Treatment -> HAQ (haq).

itreat_switch
How should the relationship between treatment and switching to a new treatment during the initial treatment phase (i.e., first 6 months) be modeled. Options, which are equivalent to S1-S6 in the documentation are S1: Treatment -> ACR -> Switch (acr-switch), S2: Treatment -> ACR -> DAS28 -> Switch (acr-das28-switch), S3: Treatment -> ACR -> SDAI -> Switch (acr-sdai-switch), S4: Treatment -> ACR -> CDAI -> Switch (acr-cdai-switch), S5: Treatment -> DAS28 -> Switch (das28-switch), S6: Treatment -> ACR -> EULAR -> Switch (acr-eular-switch).

si_dist
Survival distribution for serious infections. Currently rate is assumed to be constant so an exponential distribution is used.

max_months
Maximum number of months to run the model for. Default is NULL which implies that the model is simulated over each patient's lifetime.

input_data
List of input data. Required inputs are haq0, age, male, x.mort, and x.ttd as generated from input_data.

pars
List of parameters. Required parameters are rebound, acr1, acr2, acr2eular, eular2haq, haq.lprog.therapy, haq.lprog.age, logor.mort, mort.loghr.haqdif, ttsi, ttd.eular.mod, ttd.eular.good, and lt as generated from sample_pars. Additionally, if cdmards_prog is equal to "lcgm", then haq.lcgm must be included.

haq_prog_model
Model used to simulate the progression of HAQ. Options are "lcgm" and "linear", with "lcgm" as the default. If "lcgm" is chosen, then a latent class growth model is used for cDMARDs and NBT but a constant annual rate is is assumed for all other therapies; otherwise a constant linear HAQ progression is assumed for all therapies including cDMARDs and NBT.

ttd_dist
Survival distribution for treatment duration.

cdmards_ind
Index for cDMARDs.

nbt_ind
Index for the non-biologic (i.e. a term used to define a selection of treatments clinicians use after the last biologic in a treatment sequence).

check
Should the function check parameters and input data passed to the model? Default is TRUE.

Value

A data.table with the following columns:

sim

Simulation number. Indexed from 1 to S where S is the number of randomly sampled parameter sets (e.g. n from sample_pars).

id

ID number. Indexed from 1 to N where N is the number of simulated patients (e.g. n from sample_pats).

month

Month since a simulated patient began the first therapy in a treatment sequence.

therapy

Index of therapy used. Given J total therapies, the first J - 1 therapies match the indices from arminds. The final therapy is always equal to the index of the non-biologic therapy (nbt_ind).

therapy_seq

Number of therapies in a sequence. First therapy equal to 1, second therapy equal to 2, …

therapy_cycle

Number of model cycles since a patient began taking a given therapy in a treatment sequence. therapy_cycle = 1 during the initial 6-month treatment period.

death

Equal to 1 if patient died during the model cycle and 0 otherwise.

age

Age of patient which increases with the model cycles.

ttd

Time to treatment discontinuation. Measured in terms of model cycles (e.g. ttd = 2 if therapy will discontinue in 1 year given 6-months cycles). ttd is measured at the end of each cycle. Patients switch therapies during the cycle in which ttd becomes negative and HAQ rebounds during the cycle.

acr

Simulated ACR response during the initial 6-month period for a new therapy. Constant within therapy. Categories are 0 (ACR < 20), 1 (ACR 20-50), 2 (ACR 50-70), and 3 (ACR 70+).

eular

Simulated EULAR response during the initial 6-month period for a new therapy. Constant within therapy. Categories are 0 (no EULAR response), 1 (moderate EULAR response), and 2 (good EULAR response).

haq

HAQ score. Restricted to range between 0 and 3.

ttsi

Time to serious infection. Like ttd, measured in terms of model cycles. ttsi is measured at the end of each cycle.

si

Equal to 1 if treatment discontinuation was caused by a serious infection and 0 otherwise.

yrlen

Length of a model cycle in years. Equal to 0.5 given 6-month cycles.